back/requirements.txt
djangorestframework==3.15.2
mysqlclient==2.2.5
Django==5.1.2
python-dotenv==1.0.1
